Wire up the form to edit geometry mode, and move the delete button there
