+++ src/lib/sidebar/EditFeatureForm.svelte
@@ -1,0 +1,50 @@
+<script lang="ts" generics="F, S">
+  import { deleteIntervention, mode } from "$lib/draw/stores";
+  import { ErrorMessage, WarningButton } from "govuk-svelte";
+  import { type Config } from "$lib/config";
+  import type { FeatureWithID, Schemes } from "$lib/draw/types";
+  import type { Writable } from "svelte/store";
+
+  export let cfg: Config<F, S>;
+  export let gjSchemes: Writable<Schemes<F, S>>;
+  export let id: number;
+
+  let feature = $gjSchemes.features.find((f) => f.id == id)!;
+
+  // Because of how properties are bound individually, updates to $gjScheme aren't seen. Force them.
+  function featureUpdated(feature: FeatureWithID<F>) {
+    $gjSchemes = $gjSchemes;
+  }
+  $: featureUpdated(feature);
+
+  function onKeydown(e: KeyboardEvent) {
+    if (e.key == "Delete") {
+      const tag = (e.target as HTMLElement).tagName;
+      // Let the delete key work in forms
+      if (tag == "INPUT" || tag == "TEXTAREA") {
+        return;
+      }
+      e.stopPropagation();
+
+      deleteIntervention(gjSchemes, id);
+    }
+  }
+</script>
+
+<svelte:window on:keydown={onKeydown} />
+
+<h2>Editing {cfg.interventionName(feature)}</h2>
+
+<WarningButton on:click={() => deleteIntervention(gjSchemes, id)}>
+  Delete
+</WarningButton>
+
+<ErrorMessage errorMessage={cfg.interventionWarning(feature)} />
+
+<svelte:component
+  this={cfg.editFeatureForm}
+  {cfg}
+  {gjSchemes}
+  {id}
+  bind:props={feature.properties}
+/>
