+++ src/lib/sidebar/index.ts
@@ -1,3 +1,3 @@
-export { default as EditForm } from "./EditForm.svelte";
+export { default as EditFeatureForm } from "./EditFeatureForm.svelte";
 export { default as ListMode } from "./ListMode.svelte";
 export { default as PerSchemeControls } from "./PerSchemeControls.svelte";
