+++ src/routes/+page.svelte
@@ -16,7 +16,7 @@
     NewFeatureForm,
   } from "$lib/draw";
   import { mode, emptySchemes } from "$lib/draw/stores";
-  import { ListMode, EditForm } from "$lib/sidebar";
+  import { ListMode, EditFeatureForm } from "$lib/sidebar";
   import ExampleFeatureForm from "./ExampleFeatureForm.svelte";
   import ExampleSchemeForm from "./ExampleSchemeForm.svelte";
   import { writable } from "svelte/store";
@@ -146,10 +146,10 @@
   <div class="sidebar">
     {#if $mode.mode == "list"}
       <ListMode {cfg} {gjSchemes} />
-    {:else if $mode.mode == "edit-form"}
-      <EditForm {cfg} {gjSchemes} id={$mode.id} />
     {:else if $mode.mode == "new-point" || $mode.mode == "new-freehand-polygon" || $mode.mode == "new-snapped-polygon" || $mode.mode == "new-route"}
       <NewFeatureForm {cfg} {gjSchemes} />
+    {:else if $mode.mode == "edit-geometry"}
+      <EditFeatureForm {cfg} {gjSchemes} id={$mode.id} />
     {/if}
   </div>
   <div class="map">
